Add Navigation
Added Navigation and made Architecture
