+++ App.tsx
@@ -1,109 +1,61 @@
-import React from 'react';
-import { Image, SafeAreaView, StyleSheet, Text, View } from 'react-native';
-import { SimpleLineIcons } from '@expo/vector-icons';
+import * as React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { LoginScreen } from './src/features/login';
+import { ProfileScreen } from './src/features/profile';
+import { SettingsScreen } from './src/features/settings';
+import { SwiperScreen } from './src/features/swiper';
 
-import { Button } from './src/components/button';
-import { InlineButton } from './src/components/inline-button';
-import { Input } from './src/components/input';
+import { TabBarIcon } from './src/components/tab-bar-icon';
 
-import loginBackground from './assets/login-background.jpg';
+const Tab = createBottomTabNavigator();
 
 export default function App() {
   return (
-    <SafeAreaView style={styles.container}>
-      <Image
-        source={loginBackground}
-        resizeMode="cover"
-        style={styles.background}
-      />
-      <Text style={styles.title}>Welcome</Text>
-      <View style={styles.login}>
-        <View style={styles.header}>
-          <Text style={styles.headerTitle}>Login</Text>
-          <InlineButton title="SIGN UP" />
-        </View>
-        <Input icon="user" placeholder="Login" />
-        <Input icon="lock" placeholder="Password" secure />
-        <InlineButton title="Forget Password?" />
-        <Button title="Login" customStyle={styles.button} />
-      </View>
-      <View style={styles.footer}>
-        <Text>Login with:</Text>
-        <SimpleLineIcons
-          style={styles.social}
-          name="social-facebook"
-          size={24}
-          color="black"
+    <NavigationContainer>
+      <Tab.Navigator
+        screenOptions={({ route }) => ({
+          tabBarShowLabel: false,
+          headerShown: false,
+        })}
+      >
+        <Tab.Screen
+          name="Login"
+          component={LoginScreen}
+          options={{
+            tabBarIcon: ({ focused }) => (
+              <TabBarIcon name="login" focused={focused} />
+            ),
+          }}
         />
-        <SimpleLineIcons
-          style={styles.social}
-          name="social-google"
-          size={24}
-          color="black"
+        <Tab.Screen
+          name="Profile"
+          component={ProfileScreen}
+          options={{
+            tabBarIcon: ({ focused }) => (
+              <TabBarIcon name="ghost" focused={focused} />
+            ),
+          }}
         />
-      </View>
-    </SafeAreaView>
+        <Tab.Screen
+          name="Settings"
+          component={SettingsScreen}
+          options={{
+            tabBarIcon: ({ focused }) => (
+              <TabBarIcon name="settings" focused={focused} />
+            ),
+          }}
+        />
+        <Tab.Screen
+          name="Swiper"
+          component={SwiperScreen}
+          options={{
+            tabBarIcon: ({ focused }) => (
+              <TabBarIcon name="fire" focused={focused} />
+            ),
+          }}
+        />
+      </Tab.Navigator>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'space-evenly',
-  },
-  title: {
-    fontSize: 36,
-    fontWeight: 'bold',
-    backgroundColor: '#ffffffa7',
-    paddingHorizontal: 20,
-    paddingVertical: 4,
-  },
-  background: {
-    flex: 1,
-    width: '100%',
-    height: '50%',
-    position: 'absolute',
-    top: 0,
-    zIndex: -1,
-  },
-  login: {
-    width: '90%',
-    alignItems: 'center',
-    backgroundColor: '#fff',
-    borderRadius: 16,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 3,
-    },
-    shadowOpacity: 0.27,
-    shadowRadius: 4.65,
-
-    elevation: 6,
-    paddingHorizontal: 20,
-    paddingVertical: 40,
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    width: '100%',
-    marginBottom: 30,
-  },
-  headerTitle: {
-    textTransform: 'uppercase',
-    fontSize: 24,
-    fontWeight: 'bold',
-  },
-  button: {
-    position: 'absolute',
-    bottom: -20,
-  },
-  footer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  social: {
-    marginHorizontal: 10,
-  },
-});
