+++ src/features/settings/index.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { Text, View } from 'react-native';
+
+export const SettingsScreen = () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Settings!</Text>
+    </View>
+  );
+};
