+++ src/features/swiper/index.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { Text, View } from 'react-native';
+
+export const SwiperScreen = () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Swiper!</Text>
+    </View>
+  );
+};
